src/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "toolbox_continu_inzicht"
description = "Toolbox voor waterkeringbeheerders om risico's in kaart te brengen"
readme = "README.md"
authors = [
    { name = "David Haasnoot", email = "d.haasnoot@hkv.nl" },
    { name = "Jerom Aerts", email = "j.aerts@hkv.nl" },
    { name = "David Haasnoot", email = "d.haasnoot@hkv.nl" },
    { name = "Guy Dupuits", email = "dupuits@hkv.nl" },
    { name = "Bart Thonus", email = "b.thonus@hkv.nl" },
    { name = "Rolf Waterman", email = "r.watermans@hkv.nl" },
    { name = "Marit Zethof", email = "m.zethof@hkv.nl" },

]

license = { text = "GPL-3.0-or-later" }

requires-python = ">=3.9"
dependencies = [
                "fiona",
                "geopandas",
                "matplotlib",
                "netcdf4>=1.7.1",
                "openpyxl",
                "openturns>=1.23",
                "pandas",
                "psycopg2",
                "pyarrow",
                "pydantic>=2",
                "pyogrio",
                "python-dotenv>=1.0.1",
                "requests",
                "shapely",
                "httpx",
                "sqlalchemy",
                "tqdm>=4.67.1",
                "probabilistic_piping==0.0.5",
                "pydra_core==0.0.9",
                "pyyaml",
                "xarray",
]

dynamic = ["version"]

[tool.hatch.version]
path = "toolbox_continu_inzicht/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["toolbox_continu_inzicht"]

[project.urls]
Documentation = "https://continu-inzicht.github.io/toolbox-continu-inzicht/"
Source = "https://github.com/continu-inzicht/toolbox-continu-inzicht"
